flesh out hook support into easybuild.tools.hooks module
