+++ test/framework/filetools.py
@@ -1542,74 +1542,6 @@ class FileToolsTest(EnhancedTestCase):
         regex = re.compile('^--- .*/foo\s*\n\+\+\+ .*/bar\s*$', re.M)
         self.assertTrue(regex.search(res), "Pattern '%s' found in: %s" % (regex.pattern, res))
 
-    def test_hooks(self):
-        """Test for functions that support use of hooks."""
-        test_hooks_pymod = os.path.join(self.test_prefix, 'test_hooks.py')
-        test_hooks_pymod_txt = '\n'.join([
-            'def start_hook():',
-            '    print("running start hook")',
-            '',
-            'def foo():',
-            '    print("running foo helper method")',
-            '',
-            'def post_configure_hook(self):',
-            '    print("running post-configure hook")',
-            '    foo()',
-            '',
-            'def pre_install_hook(self):',
-            '    print("running pre-install hook")',
-        ])
-        ft.write_file(test_hooks_pymod, test_hooks_pymod_txt)
-
-        hooks = ft.load_hooks(test_hooks_pymod)
-        self.assertEqual(len(hooks), 3)
-        self.assertEqual(sorted(h.__name__ for h in hooks), ['post_configure_hook', 'pre_install_hook', 'start_hook'])
-        self.assertTrue(all(callable(h) for h in hooks))
-
-        post_configure_hook = [h for h in hooks if h.__name__ == 'post_configure_hook'][0]
-        pre_install_hook = [h for h in hooks if h.__name__ == 'pre_install_hook'][0]
-        start_hook = [h for h in hooks if h.__name__ == 'start_hook'][0]
-
-        self.assertEqual(ft.find_hook('configure', hooks), None)
-        self.assertEqual(ft.find_hook('configure', hooks, pre_step_hook=True), None)
-        self.assertEqual(ft.find_hook('configure', hooks, post_step_hook=True), post_configure_hook)
-
-        self.assertEqual(ft.find_hook('install', hooks), None)
-        self.assertEqual(ft.find_hook('install', hooks, pre_step_hook=True), pre_install_hook)
-        self.assertEqual(ft.find_hook('install', hooks, post_step_hook=True), None)
-
-        self.assertEqual(ft.find_hook('build', hooks), None)
-        self.assertEqual(ft.find_hook('build', hooks, pre_step_hook=True), None)
-        self.assertEqual(ft.find_hook('build', hooks, post_step_hook=True), None)
-
-        self.assertEqual(ft.find_hook('start', hooks), start_hook)
-        self.assertEqual(ft.find_hook('start', hooks, pre_step_hook=True), None)
-        self.assertEqual(ft.find_hook('start', hooks, post_step_hook=True), None)
-
-        self.mock_stdout(True)
-        self.mock_stderr(True)
-        ft.run_hook('start', hooks)
-        ft.run_hook('configure', hooks, pre_step_hook=True, args=[None])
-        ft.run_hook('configure', hooks, post_step_hook=True, args=[None])
-        ft.run_hook('build', hooks, pre_step_hook=True, args=[None])
-        ft.run_hook('build', hooks, post_step_hook=True, args=[None])
-        ft.run_hook('install', hooks, pre_step_hook=True, args=[None])
-        ft.run_hook('install', hooks, post_step_hook=True, args=[None])
-        stdout = self.get_stdout()
-        stderr = self.get_stderr()
-        self.mock_stdout(False)
-        self.mock_stderr(False)
-
-        expected_stdout = '\n'.join([
-            "running start hook",
-            "running post-configure hook",
-            "running foo helper method",
-            "running pre-install hook",
-        ])
-
-        self.assertEqual(stdout.strip(), expected_stdout)
-        self.assertEqual(stderr, '')
-
 
 def suite():
     """ returns all the testcases in this module """
